All environment variables moved to info file This removes the ability to specify certain cli parameters such as test rail prefix and reports directory. For now, please specify these in the info file. All info files specific to your test bed should be included in the directory 'test_bed_info'. You can run the sanity script by specify the info file with -f cli parameter, (eg. 'python3 cicd_sanity -f info_file_name'). Alternatively you can give the path to the file (eg. 'python3 cicd_sanity -f /path/to/file/info_file_name'). delete_profiles.py was added to delete profiles that cicd_sanity.py creates incase the script does not make it to completion. It can be invoked by 'python3 delete_profiles.py -f info_file_name'. It can be added to run serially to main cicd sanity with a semicolon inbetween run commands. This will be changed in the future to delete by querying the rest api by name and will be automated by including in the main cicd_sanity script.

+++ tests/cicd_sanity/delete_profiles.py
@@ -31,24 +31,25 @@ equipment_id_dict = test_info.equipment_id_dict
 bearer = CloudSDK.get_bearer(cloudSDK_url, cloud_type, cloud_user, cloud_password)
 
 
-print("Switching AP's to default profile")
-for id in equipment_id_dict.values():
-    ap_profile = CloudSDK.set_ap_profile(id, 6, cloudSDK_url, bearer)
-print('Profile change successful')
-
 with open(deletion_file) as infile:
     data = json.load(infile)
 
 delete_list = data[cloudSDK_url]
 
-for x in delete_list:
-    delete_profile = CloudSDK.delete_profile(cloudSDK_url, bearer, str(x))
-    if delete_profile == "SUCCESS":
-        print("profile", x, "delete successful")
-    else:
-        print("Error deleting profile")
+if len(delete_list) > 0:
+    print("Switching AP's to default profile")
+    for id in equipment_id_dict.values():
+        ap_profile = CloudSDK.set_ap_profile(id, 6, cloudSDK_url, bearer)
+    print('Profile change successful')
 
-data[cloudSDK_url] = []
+    for x in delete_list:
+        delete_profile = CloudSDK.delete_profile(cloudSDK_url, bearer, str(x))
+        if delete_profile == "SUCCESS":
+            print("profile", x, "delete successful")
+        else:
+            print("Error deleting profile")
 
-with open(deletion_file, 'w') as outfile:
-    json.dump(data, outfile)
+    data[cloudSDK_url] = []
+
+    with open(deletion_file, 'w') as outfile:
+        json.dump(data, outfile)
